refactor: ViewPainter 구조 개선
